Removed glitch
Many thanks for this! We are back from the break. Happy to merge this ... but as an alternative. Please provide a new PR not touching the existing Apple logo and move the black one to hexcode E06E. Also there seems to be a small glitch, please fix as well. Happy to merge after. Cheers 👍 #512